""github.com/argoproj/argo-cd/util/trunc"" to repository.go
